refactor(chat handlers backend): move hendleConnect, handleDisconnect and handleMessasges to their own files
